refactor: NumberGenerator를 Bean으로 등록하도록 변경
