+++ src/main/java/racingcar/ConsoleRacingCarApplication.java
@@ -8,19 +8,18 @@ import racingcar.repository.dao.ConsoleUserDao;
 import racingcar.repository.dao.ConsoleWinnerDao;
 import racingcar.service.FindRecordService;
 import racingcar.service.MainRacingCarService;
-import racingcar.service.PlayRacingCarService;
 import racingcar.service.SaveRacingCarResultService;
+import racingcar.domain.RandomNumberGenerator;
 
 public class ConsoleRacingCarApplication {
 
     public static void main(String[] args) {
-        final PlayRacingCarService playRacingCarService = new PlayRacingCarService();
         final FindRecordService findRecordService = new FindRecordService(new ConsoleFindAllRecordDao());
         final SaveRacingCarResultService saveRacingCarResultService = new SaveRacingCarResultService(
             new ConsoleUserDao(), new ConsoleGameDao(), new ConsolePositionDao(), new ConsoleWinnerDao());
 
         final MainRacingCarService mainService = new MainRacingCarService(
-            playRacingCarService, findRecordService, saveRacingCarResultService);
+            new RandomNumberGenerator(), findRecordService, saveRacingCarResultService);
 
         final ConsoleRacingCarController consoleController = new ConsoleRacingCarController(mainService);
         consoleController.run();
