+++ src/main/java/racingcar/domain/Cars.java
@@ -3,7 +3,6 @@ package racingcar.domain;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-import racingcar.utils.NumberGenerator;
 
 public class Cars {
