+++ src/main/java/racingcar/domain/RacingCarGame.java
@@ -2,7 +2,6 @@ package racingcar.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import racingcar.utils.NumberGenerator;
 
 public class RacingCarGame {
