+++ src/main/java/racingcar/domain/RandomNumberGenerator.java
@@ -1,0 +1,12 @@
+package racingcar.domain;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class RandomNumberGenerator implements NumberGenerator {
+
+    @Override
+    public int generate() {
+        return (int)(Math.random() * 10);
+    }
+}
