--- src/main/java/racingcar/utils/RandomNumberGenerator.java
@@ -1,9 +1,0 @@
-package racingcar.utils;
-
-public class RandomNumberGenerator implements NumberGenerator {
-
-    @Override
-    public int generate() {
-        return (int) (Math.random() * 10);
-    }
-}
